requirements.txt
python-telegram-bot[job-queue]
python-dotenv
Flask
Flask-Babel
APScheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-i18n
PyYAML